Planet Generate
Creation Planet generation

+++ UTG/Planet/Objects/PlanetData.cs
@@ -9,13 +9,13 @@ namespace UTG.Planet.Objects
 {
     public class PlanetData
     {
-        public int HeliumQuantity { get; set; }
-        public int HidrogenQuantity { get; set; }
-        public int IronQuantity { get; set; }
-        public int OxygenQuantity { get; set; }
-        public int PlatiniumQuantity { get; set; }
-        public int UraniumQuantity { get; set; }
-        public int VanadiumQuantity { get; set; }
+        public double HeliumQuantity { get; set; }
+        public double HidrogenQuantity { get; set; }
+        public double IronQuantity { get; set; }
+        public double OxygenQuantity { get; set; }
+        public double PlatiniumQuantity { get; set; }
+        public double UraniumQuantity { get; set; }
+        public double VanadiumQuantity { get; set; }
         public IPlanet Type { get; set; }
         public string Name { get; set; }
         public double Size { get; set; }
